requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.1
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2

# Database
pymongo==4.6.0
motor==3.3.2

# Environment and configuration
python-dotenv==1.0.0

# Data processing
matplotlib==3.8.2
seaborn==0.13.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Jupyter notebooks (for development)
jupyter==1.0.0
ipykernel==6.27.1

# Logging and monitoring
loguru==0.7.2
